Swap the stop/copy order in the restore script
Redis will dump an RDB file on exit when it's gracefully shut down.
On shared instances this was not a problem because the restore script
was killing it manually with a KILL signal, but on dedicated instances
this is done by monit, which sends an TERM signal.

The original order of the restore script (first copy the dump.rdb then stop
redis) caused the copied dump.rdb to be replaced on the stop.

This commit ensures that we first stop redis, then copy the rdb file to
the instance.

We also changed tests to reproduce the failing scenario.

Signed-off-by: Tiago Scolari <tiago@cloudcredo.com>