Strip quotation marks from password in redis.conf
In older versions of our release we used to quote the requirepass value
inside the redis conf file. That did not cause problems at the time
because we were not reading password from that config file when binding
an instance. Instead we were reading it from file redis-server.password
which did not use quotes. In later versions we got rid of the password
file and instead read the password from the redis config directly. This
leads to problems for customers that upgraded. Bind will give them an
incorectly quoted password.

[#97889764]

Signed-off-by: Stev Witzel <switzel@pivotal.io>