Try PING rather than check port
This is related to the shared-vm instance restart bug. We intially
checked for correctness of PID by port, this proved to be flakey because
we had no reliable way to obtain the correct port by PID. We have now
implemented a PING check that verifies if we have the correct
credentials, then the server is healthy.

[#120671733]

Signed-off-by: Tom Godkin <tgodkin@pivotal.io>