Fix hanging call to httptest.Server.Close()
Sometimes httptest.Server.Close() seems to block forever. There's an
issue that mentions this problem: https://github.com/golang/go/issues/13032
That issue was supposed to be fixed but we still see this happening on
sometimes. I'm going to lower the request interval in the hope that this
gives the server enough time to shutdown.