redis-broker tests now use only ips for dedicated nodes, not hostnames
After the latest update broker.yml only allows IPv4 for dedicated node
addresses, however the broker tests were using assets with hostnames.
These hostnames needed to be unique so they can identify nodes in the
statefile, but also all need to point to localhost for the tests.

The hack/solution is to pass versions of localhost ip which compare as
unique strings, but resolve to the same IP. i.e. "127.0.0.1", "127.0.0.01",
"127.0.0.001".

The asset that now contains "127.0.0.2" uses this IP to check
the number of available_instances, but does not attempt to provision with it.

[#150790345]

Signed-off-by: Jack Newberry <jnewberry@pivotal.io>